проба V4.3

+++ src/main/java/ru/datana/siemensopc/moka/contants/EnumConnectionMoka7Type.java
@@ -9,10 +9,10 @@ import lombok.ToString;
 public enum EnumConnectionMoka7Type {
 
     // Connection types
-    PG(0x010),
-    OP(0x02),
-    S7_BASIC(0x03);
+    PG((short) 0x010),
+    OP((short) 0x02),
+    S7_BASIC((short) 0x03);
 
     @Getter
-    private final int code;
+    private final short code;
 }
